Checkpoint what I'm working on; need to go home and make dinner and sleep.
This is going to be divided into multiple files (where the huge
headings are) after more of it works; the idea is to have a generic
AST and a placeholder logic module to allow work on the filesystem
code to proceed.

currently what I've got isn't working at all though.

Rob (or Eric), if you can change the logic part to something closer to
real that would be helpful, but don't spend a long time on it; I would
rather you spent the time on the real logic.