First cut semantics.
This has been hacked up with a placeholder Value type in a few places
to avoid having to deal with the value typing problem up front.

Also it's missing the getlock/putlock statements since I'm not really
clear on what they need to do at this level.

But, Coq accepts it, so it's a good start.