pile o' new stuff.
I haven't removed Eric's lemmas and proofs (and I mostly but not
entirely updated them for the upstream changes I made) but I've added
a new set using yesterday's new and improved astprops.

This set also includes stuff about threads and machines, not just
statements and expressions, and I took a stab at proving some of the
progress cases. Although some of them are getting stuck wanting
decidable equality on types and we might have to give up on the
polymorphic environments :-(