Stop compiling vfs.v and caponfs.v.
These depend on being able to have ast-level procs whose arguments
are type "vnode", where "vnode" is currently a Coq type class.

(a) Adding a case for "vnode" to the ast-level type inductive means
putting the definition of "vnode" where ast.v can see it, which would
make a huge mess. Plus it's disgusting.

(b) Adding struct types and type classes to the ast is a big deal.

Neither of these is going to be happening this weekend.