I figured out what's up with locks and MemoryHeap.
Nothing actually prohibits creating lock values whose address
indicates a disk. Adding typing structure to eliminate ill-formed
values is expensive (it adds extra predicates ~everywhere and turns a
lot of statements into a /\ b forms that need extra manual attention);
the right fix is to not have the whichheap field in locks at all, but
I don't want to do that at this stage because it might blow up the
logic.

When this gets fixed the separate "addr" type should get nuked. It
doesn't serve any useful purpose outside of the logic, and it's not
clear that it really does inside either...